requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.34.0
pydantic==2.8.0
python-dotenv==1.0.1
google-auth==2.43.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.192.0